requirements.txt
tensorflow
tensorflow-hub==0.7.0
pillow
selenium
face_recognition
